clean up old seeder code and fix avatar upload issues
